Refactor jobs structure by consolidating task exports and removing unused files.
Introduce new knowledge tasks for generating documentation and schema metadata suggestions. Update generateReview task to utilize new reviewPrompt. Clean up imports and enhance organization within the jobs package.
