+++ frontend/packages/jobs/src/index.ts
@@ -1,20 +1,3 @@
-export type {
-  GenerateReviewPayload,
-  ReviewResponse,
-} from './tasks/review/generateReview'
-export type { PostCommentPayload } from './tasks/review/postComment'
-
-export { processGenerateReview } from './tasks/review/generateReview'
-export { processSavePullRequest } from './tasks/review/savePullRequest'
-export { processSaveReview } from './tasks/review/saveReview'
-export { postComment } from './tasks/review/postComment'
-
-export {
-  savePullRequestTask,
-  generateReviewTask,
-  saveReviewTask,
-  postCommentTask,
-} from './tasks/review'
-export { helloWorld } from './trigger/jobs'
-
-export * from './prompts'
+export * from './tasks/review'
+export * from './tasks/knowledge'
+export * from './tasks/misc'
