--- frontend/packages/jobs/src/prompts/generateDocsSuggestion/index.ts
@@ -1,2 +1,0 @@
-export * from './docsSuggestionSchema'
-export * from './generateDocsSuggestion'
