--- frontend/packages/jobs/src/prompts/generateReview/index.ts
@@ -1,6 +1,0 @@
-export {
-  SYSTEM_PROMPT,
-  USER_PROMPT,
-  reviewJsonSchema,
-  chain,
-} from './generateReview'
