--- frontend/packages/jobs/src/prompts/index.ts
@@ -1,1 +1,0 @@
-export * from './generateReview'
