+++ frontend/packages/jobs/src/tasks/knowledge/generateSchemaMeta.ts
@@ -7,8 +7,15 @@ import {
   type SchemaOverride,
   schemaOverrideSchema,
 } from '@liam-hq/db-structure'
+import { logger, task } from '@trigger.dev/sdk/v3'
 import { toJsonSchema } from '@valibot/to-json-schema'
+import { v4 as uuidv4 } from 'uuid'
 import { type InferOutput, boolean, object, parse, string } from 'valibot'
+import { SCHEMA_OVERRIDE_FILE_PATH } from '../../constants'
+import { langfuseLangchainHandler } from '../../functions/langfuseLangchainHandler'
+import { createClient } from '../../libs/supabase'
+import { fetchSchemaInfoWithOverrides } from '../../utils/schemaUtils'
+import { createKnowledgeSuggestionTask } from './createKnowledgeSuggestion'
 
 // Define evaluation schema using valibot
 const evaluationSchema = object({
@@ -17,23 +24,121 @@ const evaluationSchema = object({
   suggestedChanges: string(),
 })
 
-// Convert schemas to JSON format for LLM
-const evaluationJsonSchema = toJsonSchema(evaluationSchema)
-const schemaOverrideJsonSchema = toJsonSchema(schemaOverrideSchema)
-
 // Define type for evaluation result
 type EvaluationResult = InferOutput<typeof evaluationSchema>
 
-type GenerateSchemaMetaResult =
+export type GenerateSchemaMetaPayload = {
+  overallReviewId: number
+}
+
+export type SchemaMetaResult =
   | {
-      updateNeeded: true
+      createNeeded: true
       override: SchemaOverride
-      reasoning: string
+      projectId: number
+      pullRequestNumber: number
+      branchName: string
+      title: string
+      traceId: string
+      overallReviewId: number
+      reasoning?: string
     }
   | {
-      updateNeeded: false
-      reasoning: string
+      createNeeded: false
     }
+
+export const processGenerateSchemaMeta = async (
+  payload: GenerateSchemaMetaPayload,
+): Promise<SchemaMetaResult> => {
+  try {
+    const supabase = createClient()
+
+    // Get the overall review from the database with nested relations
+    const { data: overallReview, error } = await supabase
+      .from('OverallReview')
+      .select(`
+        *,
+        pullRequest:PullRequest(*,
+          repository:Repository(*)
+        ),
+        project:Project(*)
+      `)
+      .eq('id', payload.overallReviewId)
+      .single()
+
+    if (error || !overallReview) {
+      throw new Error(
+        `Overall review with ID ${payload.overallReviewId} not found: ${JSON.stringify(error)}`,
+      )
+    }
+
+    const { pullRequest, project } = overallReview
+    if (!pullRequest) {
+      throw new Error(
+        `Pull request not found for overall review ${payload.overallReviewId}`,
+      )
+    }
+
+    if (!project) {
+      throw new Error(
+        `Project not found for overall review ${payload.overallReviewId}`,
+      )
+    }
+
+    const { repository } = pullRequest
+    if (!repository) {
+      throw new Error(`Repository not found for pull request ${pullRequest.id}`)
+    }
+
+    const predefinedRunId = uuidv4()
+    const callbacks = [langfuseLangchainHandler]
+
+    // Fetch schema information with overrides
+    const repositoryFullName = `${repository.owner}/${repository.name}`
+    const { currentSchemaMeta, overriddenSchema } =
+      await fetchSchemaInfoWithOverrides(
+        Number(project.id),
+        overallReview.branchName,
+        repositoryFullName,
+        Number(repository.installationId),
+      )
+
+    const schemaMetaResult = await generateSchemaMeta(
+      overallReview.reviewComment || '',
+      callbacks,
+      currentSchemaMeta,
+      predefinedRunId,
+      overriddenSchema,
+    )
+
+    // If no update is needed, return early with createNeeded: false
+    if (!schemaMetaResult.updateNeeded) {
+      return {
+        createNeeded: false,
+      }
+    }
+
+    // Return the schema meta along with information needed for createKnowledgeSuggestionTask
+    return {
+      createNeeded: true,
+      override: schemaMetaResult.override,
+      projectId: project.id,
+      pullRequestNumber: Number(pullRequest.pullNumber), // Convert bigint to number
+      branchName: overallReview.branchName, // Get branchName from overallReview
+      title: `Schema meta update from PR #${Number(pullRequest.pullNumber)}`,
+      traceId: predefinedRunId,
+      reasoning: schemaMetaResult.reasoning,
+      overallReviewId: payload.overallReviewId,
+    }
+  } catch (error) {
+    console.error('Error generating schema meta:', error)
+    throw error
+  }
+}
+
+// Convert schemas to JSON format for LLM
+const evaluationJsonSchema = toJsonSchema(evaluationSchema)
+const schemaOverrideJsonSchema = toJsonSchema(schemaOverrideSchema)
 
 // Step 1: Evaluation template to determine if updates are needed
 const EVALUATION_TEMPLATE = ChatPromptTemplate.fromTemplate(`
@@ -305,3 +410,45 @@ export const generateSchemaMeta = async (
     throw error
   }
 }
+
+export type GenerateSchemaMetaResult =
+  | {
+      updateNeeded: true
+      override: SchemaOverride
+      reasoning: string
+    }
+  | {
+      updateNeeded: false
+      reasoning: string
+    }
+
+export const generateSchemaMetaSuggestionTask = task({
+  id: 'generate-schema-meta-suggestion',
+  run: async (payload: GenerateSchemaMetaPayload) => {
+    logger.log('Executing schema meta suggestion task:', { payload })
+    const result = await processGenerateSchemaMeta(payload)
+    logger.info('Generated schema meta suggestion:', { result })
+
+    if (result.createNeeded) {
+      // Create a knowledge suggestion with the schema meta using the returned information
+      await createKnowledgeSuggestionTask.trigger({
+        projectId: result.projectId,
+        type: 'SCHEMA',
+        title: result.title,
+        path: SCHEMA_OVERRIDE_FILE_PATH,
+        content: JSON.stringify(result.override, null, 2),
+        branch: result.branchName,
+        traceId: result.traceId,
+        reasoning: result.reasoning || '',
+        overallReviewId: result.overallReviewId,
+      })
+      logger.info('Knowledge suggestion creation triggered')
+    } else {
+      logger.info(
+        'No schema meta update needed, skipping knowledge suggestion creation',
+      )
+    }
+
+    return { result }
+  },
+})
