+++ frontend/packages/jobs/src/tasks/knowledge/index.ts
@@ -1,0 +1,3 @@
+export * from './createKnowledgeSuggestion'
+export * from './generateDocsSuggestion'
+export * from './generateSchemaMeta'
