+++ frontend/packages/jobs/src/tasks/misc/index.ts
@@ -1,0 +1,2 @@
+export * from './helloWorld'
+export * from './sentryErrorTest'
