--- frontend/packages/jobs/src/trigger/helloworld.ts
@@ -1,9 +1,0 @@
-import { logger, task } from '@trigger.dev/sdk/v3'
-
-export const helloWorldTask = task({
-  id: 'helloworld',
-  run: async (payload: { name: string }) => {
-    logger.log('Executing Hello World task:', { payload })
-    return `Hello ${payload.name}`
-  },
-})
